// Application.RMGUI/ClientApp/src/app/models/AssignPanel.ts
import { Time } from '@angular/common';

export class AssignPanel {
 
  assigning_panel_code: string;
  assign_panel_name: string;
  applicant_name: string;
  type_of_assessment: string;
  time_slot: Time;
  assign_panel_status: string;
  assignpanel_start_date: string;
  assignpanel_end_date: string;
}